+++ docs/classes.rst
@@ -4,7 +4,7 @@ Classes
 .. toctree::
    :maxdepth: 1
 
-   pyfar.classes_audio
-   pyfar.classes_filter
-   pyfar.classes_coordinates
-   pyfar.classes_orientations
+   classes/pyfar.audio
+   classes/pyfar.filter
+   classes/pyfar.coordinates
+   classes/pyfar.orientations
