+++ docs/classes/pyfar.audio.rst
