+++ docs/classes/pyfar.coordinates.rst
