+++ docs/classes/pyfar.filter.rst
