+++ docs/concepts/pyfar.audio.rst
@@ -1,0 +1,4 @@
+Audio
+-----------------------------------------------
+
+.. automodule:: pyfar.classes.audio_concepts
