+++ docs/concepts/pyfar.coordinates.rst
@@ -1,0 +1,4 @@
+Coordinates
+-----------------------------------------------
+
+.. automodule:: pyfar.classes.coordinates_concepts
