+++ docs/concepts/pyfar.filter.rst
