+++ docs/concepts/resources/coordinate_systems.ai
