+++ docs/concepts/resources/coordinate_systems.png
