+++ docs/modules/pyfar.dsp.filter.rst
