+++ docs/modules/pyfar.io.rst
