+++ docs/modules/pyfar.plot.rst
