+++ docs/modules/pyfar.samplings.rst
