+++ docs/modules/pyfar.signals.rst
