+++ pyfar/classes/filter.py
@@ -1,7 +1,7 @@
 """
 The following documents the pyfar filter classes. More details and
 background is given in the
-:py:mod:`Filter concepts <pyfar.classes.filter_concepts>`
+:py:mod:`filter concepts <pyfar.classes.filter_concepts>`.
 """
 import deepdiff
 import warnings
